terncli: split subcommands into command files
