+++ go/terncli/cmd/cdcreq.go
@@ -1,0 +1,55 @@
+// Copyright 2025 XTX Markets Technologies Limited
+//
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+package cmd
+
+import (
+	"encoding/json"
+	"flag"
+	"fmt"
+	"github.com/XTXMarkets/ternfs/go/msgs"
+	"os"
+)
+
+func NewCdcReq() SpecWithClient {
+	cdcReqCmd := flag.NewFlagSet("cdc-req", flag.ExitOnError)
+	cdcReqKind := cdcReqCmd.String("kind", "", "")
+	cdcReqReq := cdcReqCmd.String("req", "", "Request body, in JSON")
+	cdcReqRun := func(runtime RuntimeWithClient) {
+		l := runtime.Log
+		req, resp, err := msgs.MkCDCMessage(*cdcReqKind)
+		if err != nil {
+			panic(err)
+		}
+		if err := json.Unmarshal([]byte(*cdcReqReq), &req); err != nil {
+			panic(fmt.Errorf("could not decode cdc req: %w", err))
+		}
+		fmt.Printf("Will send this CDC request: %T %+v\n", req, req)
+		for {
+			var action string
+			fmt.Printf("Proceed? y/n ")
+			fmt.Scanln(&action)
+			if action == "y" {
+				break
+			}
+			if action == "n" {
+				fmt.Printf("BYE\n")
+				os.Exit(0)
+			}
+		}
+		if err := runtime.Client.CDCRequest(l, req, resp); err != nil {
+			panic(err)
+		}
+		out, err := json.MarshalIndent(resp, "", "  ")
+		if err != nil {
+			panic(fmt.Errorf("could not encode response %+v to json: %w", resp, err))
+		}
+		os.Stdout.Write(out)
+		fmt.Println()
+	}
+	return SpecWithClient{
+		Flags: cdcReqCmd,
+		Run:   cdcReqRun,
+	}
+}
