+++ go/terncli/cmd/defragspans.go
@@ -1,0 +1,35 @@
+// Copyright 2025 XTX Markets Technologies Limited
+//
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+package cmd
+
+import (
+	"flag"
+	"github.com/XTXMarkets/ternfs/go/cleanup"
+	"github.com/XTXMarkets/ternfs/go/client"
+	"github.com/XTXMarkets/ternfs/go/core/bufpool"
+	"github.com/XTXMarkets/ternfs/go/core/log"
+)
+
+func NewDefragSpans() SpecWithClient {
+	defragSpansCmd := flag.NewFlagSet("defrag-spans", flag.ExitOnError)
+	defragSpansPath := defragSpansCmd.String("path", "", "The directory or file to defrag")
+	defragSpansRun := func(runtime RuntimeWithClient) {
+		l := runtime.Log
+		c := runtime.Client
+		dirInfoCache := client.NewDirInfoCache()
+		bufPool := bufpool.NewBufPool()
+		stats := &cleanup.DefragSpansStats{}
+		alert := l.NewNCAlert(0)
+		alert.SetAppType(log.XMON_NEVER)
+		if err := cleanup.DefragSpans(l, c, bufPool, dirInfoCache, stats, alert, *defragSpansPath); err != nil {
+			panic(err)
+		}
+		l.Info("defrag stats: %+v", stats)
+	}
+	return SpecWithClient{
+		Flags: defragSpansCmd,
+		Run:   defragSpansRun,
+	}
+}
