+++ go/terncli/cmd/testblockwrite.go
@@ -1,0 +1,84 @@
+// Copyright 2025 XTX Markets Technologies Limited
+//
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+package cmd
+
+import (
+	"bytes"
+	"flag"
+	"fmt"
+	"github.com/XTXMarkets/ternfs/go/client"
+	"github.com/XTXMarkets/ternfs/go/msgs"
+	"net"
+	"strconv"
+	"strings"
+	"sync"
+	"time"
+)
+
+func NewTestBlockWrite() Spec {
+	testBlockWriteCmd := flag.NewFlagSet("test-block-write", flag.ExitOnError)
+	testBlockWriteBlockService := testBlockWriteCmd.String("bs", "", "Block service. If comma-separated, they'll be written in parallel to the specified ones.")
+	testBlockWriteSize := testBlockWriteCmd.Uint("size", 0, "Size (must fit in u32)")
+	testBlockWriteRun := func(runtime Runtime) {
+		l := runtime.Log
+		registryAddress := runtime.RegistryAddress
+		resp, err := client.RegistryRequest(l, nil, *registryAddress, &msgs.ChangedBlockServicesReq{})
+		if err != nil {
+			panic(err)
+		}
+		blockServices := resp.(*msgs.ChangedBlockServicesResp)
+		bsInfos := []msgs.FullBlockServiceInfo{}
+		for _, str := range strings.Split(*testBlockWriteBlockService, ",") {
+			bsId, err := strconv.ParseUint(str, 0, 64)
+			if err != nil {
+				panic(err)
+			}
+			found := false
+			for _, bsInfo := range blockServices.BlockServices {
+				if bsInfo.Id == msgs.BlockServiceId(bsId) {
+					bsInfos = append(bsInfos, bsInfo)
+					found = true
+					break
+				}
+			}
+			if !found {
+				panic(fmt.Errorf("could not find block service %q", str))
+			}
+		}
+		conns := make([]*net.TCPConn, len(bsInfos))
+		for i := 0; i < len(conns); i++ {
+			conn, err := client.BlockServiceConnection(l, bsInfos[i].Addrs)
+			if err != nil {
+				panic(err)
+			}
+			conns[i] = conn
+		}
+		contents := make([]byte, *testBlockWriteSize)
+		var wait sync.WaitGroup
+		wait.Add(len(conns))
+		t := time.Now()
+		for i := 0; i < len(conns); i++ {
+			conn := conns[i]
+			bsId := bsInfos[i].Id
+			go func() {
+				thisErr := client.TestWrite(l, conn, bsId, bytes.NewReader(contents), uint64(len(contents)))
+				if thisErr != nil {
+					err = thisErr
+				}
+				wait.Done()
+			}()
+		}
+		wait.Wait()
+		elapsed := time.Since(t)
+		if err != nil {
+			panic(err)
+		}
+		l.Info("writing %v bytes to %v block services took %v (%fGB/s)", *testBlockWriteSize, len(conns), time.Since(t), (float64(*testBlockWriteSize*uint(len(conns)))/1e9)/elapsed.Seconds())
+	}
+	return Spec{
+		Flags: testBlockWriteCmd,
+		Run:   testBlockWriteRun,
+	}
+}
